Refactor NoteCard and SearchBar components with Vuetify and improved features
